Made it so System message is stored in state
This is because previously it was only added to the first request to OpenAI
Now every time we send a request the system message is included and not lost
